improvements to update script

+++ update.sh
@@ -1,6 +1,9 @@
-#!/bin/sh
-
+#!/bin/bash
+current_git=$(git rev-parse HEAD)
 git pull
-docker build -t ffufme .
-docker kill $(docker ps -q)
-docker run -d -p 80:80 ffufme
+now_git=$(git rev-parse HEAD)
+if [[ "$current_git" != "$now_git" ]]; then
+    docker build -t ffufme .
+    docker kill $(docker ps -q)
+    docker run -d -p 80:80 ffufme
+fi
